generic-amd64-fde: Enable signing by a separate SIGN variable
Right now the signing is enabled whenever the configuration is in place
(specifically the SIGN_API variable is checked to be nonempty). This patch
introduces a new SIGN variable that must be set to true to trigger signing
so that configuration can still be in place even for unsigned images.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>