linux-yocto: generic-aarch64: copy defconfig from debian
Install a config for generic-aarch64, in kind with the generic-amd64
config copied from the genericx86-64-ext config. However, copy the
config directly from the Debian Bullseye linux-config-5.10 (5.10.46-4)
package, as we don't already have a config in use for this architecture.

Also save this as a defconfig.

Changelog-entry: linux-yocto: generic-aarch64: copy defconfig from debian
Signed-off-by: Joseph Kogut <joseph@balena.io>