Use a 2nd stage bootloader to unlock LUKS partitions instead of GRUB
GRUB can not use the TPM easily to unlock the volumes and find the kernel
on an encrypted partition. Instead, we choose to store a linux kernel
and use it as 2nd stage bootloader to unlock the partition, load the actual
kernel and kexec into it. This should eventually be replaced by a proper
2nd stage bootloader that is being worked on.

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>