generic-amd64: make flasher image default artifact
We already process and upload both flasher and raw images that are
explicitly labeled, but for legacy reasons, the unlabeled image is also
used as the default. Upload the flasher image as the default/unlabeled
artifact to remain consistent with the genericx86-64-ext DT provided in
balena-intel.

Changelog-entry: Default to flasher image for generic-amd64
Signed-off-by: Joseph Kogut <joseph@balena.io>