Disable AppArmor in generic-aarch64
This aligns the generic-aarch64 kernel to the generic-amd64 kernel with
regards to AppArmor configuration.

Initial motivation for this change comes balenaEngine: newer releases of
Moby expect the AppArmor userland tools to be installed if AppArmor is
enabled on the kernel. Given that balenaOS doesn't distribute the
AppArmor tools, our aarch64 tests running on CI were failing for new
Engine releases that incorporate those Moby changes.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Changelog-entry: Disable AppArmor in aarch64-generic, to align it with amd64-generic