balena-image.inc: remove IMAGE_FSTYPES override
At this moment meta-balena-raspberrypi overrides IMAGE_FSTYPES, which
breaks the build of signed images, because we have introduced a separate
image type for that use case. The override also seems unnecessary,
as it sets the value to the exact same thing that meta-balena already
did.

This patch removes the override and makes the devices use
the definition from meta-balena, which does account for the signature
split.

Change-type: patch
Changelog-entry: Unblock builds for secure boot enabled device types
Signed-off-by: Michal Toman <michalt@balena.io>