linux: Make UAS driver built in
Until now the assumption was that USB storage devices that support UAS
mode can fall back to using usb-storage, but we have found a few
USB 3.2 drives that will only work with UAS. Such drives don't
work in the initramfs and can't be used for running flasher.

This patch makes the UAS driver built in.

Changelog-entry: Make UAS driver built in
Signed-off-by: Michal Toman <michalt@balena.io>