Blacklist nouveau and nvidiafb on secure boot device types
In general, the recommended approach for using the out-of-tree NVIDIA
drivers is to unload nouveau and nvidiafb at runtime, right before
the NVIDIA driver is loaded. Certain cards do not like this approach
and won't work properly if nouveau is loaded at all.

Since the primary driver for creating separate secure boot device
types is to use secure boot in combination with NVIDIA GPUs
and the out-of-tree driver, we can safely blacklist nouveau
and nvidiafb at build time. This makes it easier to load the NVIDIA
driver from the application and reduces the risk of a double-reset
putting the card into an undefined state.

Changelog-entry: Blacklist nouveau and nvidiafb on secure boot device types
Signed-off-by: Michal Toman <michalt@balena.io>